TimerSet now creates ScheduledTasks.
