+++ lib/concurrent/executor/timer_set.rb
@@ -1,11 +1,8 @@
-require 'concurrent/errors'
-require 'concurrent/ivar'
+require 'concurrent/scheduled_task'
 require 'concurrent/atomic/event'
 require 'concurrent/collection/priority_queue'
-require 'concurrent/executor/executor'
 require 'concurrent/executor/executor_service'
 require 'concurrent/executor/single_thread_executor'
-require 'concurrent/utility/monotonic_time'
 
 module Concurrent
 
@@ -15,97 +12,6 @@ module Concurrent
   #
   # @!macro monotonic_clock_warning
   class TimerSet < RubyExecutorService
-
-    # An `IVar` representing a tasked queued for execution in a `TimerSet`.
-    class Task < Concurrent::IVar
-      include Comparable
-
-      attr_reader :executor
-
-      def initialize(parent, delay, args, task, opts = {})
-        super(IVar::NO_VALUE, opts, &nil)
-        synchronize do
-          ns_set_delay_and_time!(delay)
-          @parent = parent
-          @args = args
-          @task = task
-          @executor = Executor.executor_from_options(opts) || Concurrent.global_io_executor
-          self.observers = CopyOnNotifyObserverSet.new
-        end
-      end
-
-      def original_delay
-        synchronize { @delay }
-      end
-
-      def schedule_time
-        synchronize { @time }
-      end
-
-      def <=>(other)
-        self.schedule_time <=> other.schedule_time
-      end
-
-      # Has the task been cancelled?
-      #
-      # @return [Boolean] true if the task is in the given state else false
-      def cancelled?
-        synchronize { ns_check_state?(:cancelled) }
-      end
-
-      # In the task execution in progress?
-      #
-      # @return [Boolean] true if the task is in the given state else false
-      def processing?
-        synchronize { ns_check_state?(:processing) }
-      end
-
-      # Cancel this task and prevent it from executing. A task can only be
-      # cancelled if it is pending or unscheduled.
-      #
-      # @return [Boolean] true if task execution is successfully cancelled
-      #   else false
-      def cancel
-        if compare_and_set_state(:cancelled, :pending, :unscheduled)
-          complete(false, nil, CancelledOperationError.new)
-          # To avoid deadlocks this call must occur outside of #synchronize
-          # Changing the state above should prevent redundant calls
-          @parent.send(:remove_task, self)
-        else
-          false
-        end
-      end
-
-      def reset
-        synchronize{ ns_reschedule(@delay) }
-      end
-
-      def reschedule(delay)
-        synchronize{ ns_reschedule(delay) }
-      end
-
-      # @!visibility private
-      def process_task
-        safe_execute(@task, @args)
-      end
-
-      protected :set, :try_set, :fail, :complete
-
-      protected
-
-      def ns_set_delay_and_time!(delay)
-        @delay = TimerSet.calculate_delay!(delay)
-        @time = Concurrent.monotonic_time + @delay
-      end
-
-      def ns_reschedule(delay, fail_if_cannot_remove = true)
-        return false unless ns_check_state?(:pending)
-        ns_set_delay_and_time!(delay)
-        removed = @parent.send(:remove_task, self)
-        return false if fail_if_cannot_remove && !removed
-        @parent.send(:post_task, self)
-      end
-    end
 
     # Create a new set of timed tasks.
     #
@@ -137,9 +43,14 @@ module Concurrent
     # @!macro deprecated_scheduling_by_clock_time
     def post(delay, *args, &task)
       raise ArgumentError.new('no block given') unless block_given?
-      task = Task.new(self, delay, args, task, {executor: @task_executor}) # may raise exception
-      ok = synchronize{ ns_post_task(task) }
-      ok ? task : false
+      return false unless running?
+      opts = {
+        executor: @task_executor,
+        args: args,
+        timer_set: self
+      }
+      task = ScheduledTask.execute(delay, opts, &task) # may raise exception
+      task.unscheduled? ? false : task
     end
 
     # Begin an immediate shutdown. In-progress tasks will be allowed to
@@ -148,30 +59,6 @@ module Concurrent
     # not running.
     def kill
       shutdown
-    end
-
-    # Schedule a task to be executed after a given delay (in seconds).
-    #
-    # @param [Float] delay the number of seconds to wait for before executing the task
-    #
-    # @return [Float] the number of seconds to delay
-    #
-    # @raise [ArgumentError] if the intended execution time is not in the future
-    # @raise [ArgumentError] if no block is given
-    #
-    # @!macro deprecated_scheduling_by_clock_time
-    #
-    # @!visibility private
-    def self.calculate_delay!(delay)
-      if delay.is_a?(Time)
-        warn '[DEPRECATED] Use an interval not a clock time; schedule is now based on a monotonic clock'
-        now = Time.now
-        raise ArgumentError.new('schedule time must be in the future') if delay <= now
-        delay.to_f - now.to_f
-      else
-        raise ArgumentError.new('seconds must be greater than zero') if delay.to_f < 0.0
-        delay.to_f
-      end
     end
 
     private :<<
